+++ src/composables/useSlotComponent.ts
@@ -1,1 +1,0 @@
-export { loadPluginVueComponent } from "../utils/compilePluginVueSfc";